+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ## Unreleased
 
+## [v0.10.1b](https://github.com/TTT-2/TTT2/tree/v0.10.1b) (2021-10-15)
+
 ### Fixed
 
 - Fixed Playermodels not correctly loading changes on game start